rename hooks utils to legacy
