+++ src/hooks/useCombobox/utils.js
@@ -2,7 +2,7 @@ import {useRef, useEffect} from 'react'
 import PropTypes from 'prop-types'
 import {isControlledProp} from '../../utils.legacy'
 import {getState} from '../../utils-ts'
-import {getInitialState as getInitialStateCommon} from '../utils'
+import {getInitialState as getInitialStateCommon} from '../utils.legacy'
 import {dropdownDefaultProps, dropdownPropTypes} from '../utils.dropdown'
 import {useIsInitialMount, useEnhancedReducer} from '../utils-ts'
 import {ControlledPropUpdatedSelectedItem} from './stateChangeTypes'
