+++ src/hooks/useMultipleSelection/index.js
@@ -1,7 +1,7 @@
 import {useRef, useEffect, useCallback, useMemo} from 'react'
 import {handleRefs, callAllEventHandlers, normalizeArrowKey} from '../../utils.legacy'
 import {useLatestRef} from '../../utils-ts'
-import {useGetterPropsCalledChecker, useControlPropsValidator} from '../utils'
+import {useGetterPropsCalledChecker, useControlPropsValidator} from '../utils.legacy'
 import {
   useControlledReducer,
   useIsInitialMount,
