+++ src/hooks/useMultipleSelection/testUtils.js
@@ -20,8 +20,8 @@ jest.mock('react', () => {
   }
 })
 
-jest.mock('../utils', () => {
-  const utils = jest.requireActual('../utils')
+jest.mock('../utils.legacy', () => {
+  const utils = jest.requireActual('../utils.legacy')
   const hooksUtils = jest.requireActual('../../utils-ts')
 
   return {
